+++ src/main/java/com/blue/tnb/service/TicketService.java
@@ -4,12 +4,10 @@ import com.blue.tnb.dto.BookResponse;
 import com.blue.tnb.dto.TicketDTO;
 import com.blue.tnb.exception.TicketExceptions.TicketNotFoundException;
 import com.blue.tnb.exception.TicketExceptions.TicketWithoutUserException;
-import com.blue.tnb.model.Play;
 import com.blue.tnb.model.Ticket;
 
 import java.text.ParseException;
 import java.util.List;
-import java.util.Optional;
 
 
 public interface TicketService {
@@ -30,7 +28,7 @@ public interface TicketService {
     Ticket updateTicket(Long id,TicketDTO ticketDTO);
     Ticket deleteTicket(Long ticketId);
 
-    Long countAvailableTicketsByPlayId(Long playId);
+    int countAvailableTicketsByPlayId(Long playId);
 
     List<TicketDTO> findAllTicketsByUserId(Long id);
     //List<TicketDTO> getAllByUser(User user);
